requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
numpy
scikit-learn==1.3.0
python-dotenv==1.0.0
aiofiles==23.2.1